comm_drv_n0183-serial: refactor: Make IO thread a separate file.
