+++ model/CMakeLists.txt
@@ -89,6 +89,7 @@ set(
   ${MODEL_HDR_DIR}/mDNS_service.h
   ${MODEL_HDR_DIR}/meteo_points.h
   ${MODEL_HDR_DIR}/multiplexer.h
+  ${MODEL_HDR_DIR}/n0183_comm_mgr.h
   ${MODEL_HDR_DIR}/nav_object_database.h
   ${MODEL_HDR_DIR}/navutil_base.h
   ${MODEL_HDR_DIR}/nmea_log.h
@@ -177,6 +178,7 @@ set(SRC
   ${MODEL_SRC_DIR}/mdns_cache.cpp
   ${MODEL_SRC_DIR}/mDNS_service.cpp
   ${MODEL_SRC_DIR}/multiplexer.cpp
+  ${MODEL_SRC_DIR}/n0183_comm_mgr.cpp
   ${MODEL_SRC_DIR}/nav_object_database.cpp
   ${MODEL_SRC_DIR}/navutil_base.cpp
   ${MODEL_SRC_DIR}/ocpn_plugin.cpp
