+++ model/include/model/n0183_comm_mgr.h
@@ -1,0 +1,66 @@
+/***************************************************************************
+ *
+ * Project:  OpenCPN
+ * Purpose:  N0183 serial communications IO thread
+ * Author:   David Register, Alec Leamas
+ *
+ ***************************************************************************
+ *   Copyright (C) 2022 by David Register, Alec Leamas                     *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ *   This program is distributed in the hope that it will be useful,       *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with this program; if not, write to the                         *
+ *   Free Software Foundation, Inc.,                                       *
+ *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301,  USA.         *
+ **************************************************************************/
+
+#ifndef _N0183_PROTOL_MGR__
+#define _N0183_PROTOL_MGR__
+
+#include "comm_buffers.h"
+
+#ifndef __ANDROID__
+#include "serial/serial.h"
+#endif
+
+class CommDriverN0183Serial;
+
+class CommDriverN0183SerialThread {
+public:
+  CommDriverN0183SerialThread(CommDriverN0183Serial* Launcher,
+                              const wxString& PortName,
+                              const wxString& strBaudRate);
+
+  ~CommDriverN0183SerialThread(void);
+  void* Entry();
+  bool SetOutMsg(const wxString& msg);
+  void OnExit(void);
+
+private:
+#ifndef __ANDROID__
+  serial::Serial m_serial;
+#endif
+  void ThreadMessage(const wxString& msg);
+  bool OpenComPortPhysical(const wxString& com_name, int baud_rate);
+  void CloseComPortPhysical();
+  ssize_t WriteComPortPhysical(const char* msg);
+
+  CommDriverN0183Serial* m_pParentDriver;
+  wxString m_PortName;
+  wxString m_FullPortName;
+
+  int m_baud;
+
+  OutputBuffer m_out_que;
+};
+
+#endif  //    _N0183_PROTOL_MGR__
