+++ model/src/comm_drv_n0183_serial.cpp
@@ -42,6 +42,7 @@
 #include "model/comm_navmsg_bus.h"
 #include "model/comm_drv_registry.h"
 #include "model/logger.h"
+#include "model/n0183_comm_mgr.h"
 #include "model/sys_events.h"
 #include "model/wait_continue.h"
 #include "observable.h"
@@ -82,39 +83,6 @@ wxEvent* CommDriverN0183SerialEvent::Clone() const {
   newevent->m_payload = this->m_payload;
   return newevent;
 };
-
-#ifndef __ANDROID__
-class CommDriverN0183SerialThread {
-public:
-  CommDriverN0183SerialThread(CommDriverN0183Serial* Launcher,
-                              const wxString& PortName,
-                              const wxString& strBaudRate);
-
-  ~CommDriverN0183SerialThread(void);
-  void* Entry();
-  bool SetOutMsg(const wxString& msg);
-  void OnExit(void);
-
-private:
-  serial::Serial m_serial;
-  void ThreadMessage(const wxString& msg);
-  bool OpenComPortPhysical(const wxString& com_name, int baud_rate);
-  void CloseComPortPhysical();
-  ssize_t WriteComPortPhysical(const std::string& msg);
-
-  CommDriverN0183Serial* m_parent_driver;
-  wxString m_port_name;
-  wxString m_full_port_name;
-
-  int m_baud;
-  size_t m_send_retries;
-
-  OutputBuffer m_out_que;
-  WaitContinue device_waiter;
-  ObsListener resume_listener;
-  ObsListener new_device_listener;
-};
-#endif
 
 CommDriverN0183Serial::CommDriverN0183Serial(const ConnectionParams* params,
                                              DriverListener& listener)
@@ -315,180 +283,3 @@ void CommDriverN0183Serial::handle_N0183_MSG(
     m_listener.Notify(std::move(msg_all));
   }
 }
-
-#ifndef __ANDROID__
-#define DS_RX_BUFFER_SIZE 4096
-
-CommDriverN0183SerialThread::CommDriverN0183SerialThread(
-    CommDriverN0183Serial* Launcher, const wxString& PortName,
-    const wxString& strBaudRate) {
-  m_parent_driver = Launcher;  // This thread's immediate "parent"
-
-  m_port_name = PortName;
-  m_full_port_name = "Serial:" + PortName;
-
-  m_baud = 4800;  // default
-  long lbaud;
-  if (strBaudRate.ToLong(&lbaud)) m_baud = (int)lbaud;
-  resume_listener.Init(SystemEvents::GetInstance().evt_resume,
-                       [&](ObservedEvt&) { device_waiter.Continue(); });
-  new_device_listener.Init(SystemEvents::GetInstance().evt_dev_change,
-                           [&](ObservedEvt&) { device_waiter.Continue(); });
-}
-
-CommDriverN0183SerialThread::~CommDriverN0183SerialThread(void) {}
-
-void CommDriverN0183SerialThread::OnExit(void) {}
-
-bool CommDriverN0183SerialThread::OpenComPortPhysical(const wxString& com_name,
-                                                      int baud_rate) {
-  try {
-    m_serial.setPort(com_name.ToStdString());
-    m_serial.setBaudrate(baud_rate);
-    m_serial.open();
-    m_serial.setTimeout(250, 250, 0, 250, 0);
-  } catch (std::exception&) {
-    //      std::cerr << "Unhandled Exception while opening serial port: " <<
-    //      e.what() << std::endl;
-  }
-  return m_serial.isOpen();
-}
-
-void CommDriverN0183SerialThread::CloseComPortPhysical() {
-  try {
-    m_serial.close();
-  } catch (std::exception&) {
-    //      std::cerr << "Unhandled Exception while closing serial port: " <<
-    //      e.what() << std::endl;
-  }
-}
-
-bool CommDriverN0183SerialThread::SetOutMsg(const wxString& msg) {
-  m_out_que.Put(msg.ToStdString());
-  return true;
-}
-
-void CommDriverN0183SerialThread::ThreadMessage(const wxString& msg) {
-  //    Signal the main program thread
-  //   OCPN_ThreadMessageEvent event(wxEVT_OCPN_THREADMSG, 0);
-  //   event.SetSString(std::string(msg.mb_str()));
-  //   if (gFrame) gFrame->GetEventHandler()->AddPendingEvent(event);
-}
-
-ssize_t CommDriverN0183SerialThread::WriteComPortPhysical(
-    const std::string& msg) {
-  if (!m_serial.isOpen()) return -1;
-
-  auto ptr = reinterpret_cast<const uint8_t*>(msg.c_str());
-  size_t written = 0;
-  int tries = 0;
-  while (written < msg.size()) {
-    int chunk_size;
-    try {
-      chunk_size = m_serial.write(ptr, msg.size() - written);
-    } catch (std::exception& e) {
-      MESSAGE_LOG << "Exception while writing to serial port: " << e.what();
-      return -1;
-    }
-    if (chunk_size < 0) return chunk_size;
-    written += chunk_size;
-    ptr += chunk_size;
-    if (tries++ > 20) {
-      MESSAGE_LOG << "Error writing data (output stalled?)";
-      return -1;
-    }
-    if (written < msg.size()) std::this_thread::sleep_for(10ms);
-    // FIXME (leamas) really?
-  }
-  return written;
-}
-
-void* CommDriverN0183SerialThread::Entry() {
-  bool not_done = true;
-  m_parent_driver->SetSecThreadActive();  // I am alive
-  wxString msg;
-  LineBuffer line_buf;
-
-  //    Request the com port from the comm manager
-  if (!OpenComPortPhysical(m_port_name, m_baud)) {
-    wxString msg("NMEA input device open failed: ");
-    msg.Append(m_port_name);
-    ThreadMessage(msg);
-    // goto thread_exit; // This means we will not be trying to connect = The
-    // device must be connected when the thread is created. Does not seem to be
-    // needed/what we want as the reconnection logic is able to pick it up
-    // whenever it actually appears (Of course given it appears with the
-    // expected device name).
-  }
-
-  //    The main loop
-  m_send_retries = 0;
-
-  while ((not_done) && (m_parent_driver->m_Thread_run_flag > 0)) {
-    if (m_parent_driver->m_Thread_run_flag == 0) break;
-
-    uint8_t next_byte = 0;
-    unsigned int newdata = 0;
-    uint8_t rdata[2000];
-
-    if (m_serial.isOpen()) {
-      try {
-        newdata = m_serial.read(rdata, sizeof(rdata));
-      } catch (std::exception& e) {
-        DEBUG_LOG << "Serial read exception: " << e.what();
-        if (10 < m_send_retries++) {
-          // We timed out waiting for the next character 10 times, let's close
-          // the port so that the reconnection logic kicks in and tries to fix
-          // our connection.
-          CloseComPortPhysical();
-          m_send_retries = 0;
-        }
-      }
-    } else {
-      // Reconnection logic. Let's try to reopen the port while waiting longer
-      // every time (until we simply keep trying every 2.5 seconds)
-      // std::cerr << "Serial port seems closed." << std::endl;
-      device_waiter.Wait(250 * m_send_retries);
-      CloseComPortPhysical();
-      if (OpenComPortPhysical(m_port_name, m_baud))
-        m_send_retries = 0;
-      else if (m_send_retries < 10)
-        m_send_retries++;
-    }
-
-    // Handle received data
-    for (unsigned i = 0; i < newdata; i++) line_buf.Put(rdata[i]);
-    while (line_buf.HasLine()) {
-      auto payload = std::make_shared<std::vector<uint8_t>>(line_buf.GetLine());
-      //    Message is ready to parse and send out
-      //    Messages may be coming in as <blah blah><lf><cr>.
-      //    One example device is KVH1000 heading sensor.
-      //    If that happens, the first character of a new captured message
-      //    will the <cr>, and we need to discard it. This is out of spec,
-      //    but we should handle it anyway
-      if (payload->at(0) == '\r') payload->erase(payload->begin());
-
-      CommDriverN0183SerialEvent Nevent(wxEVT_COMMDRIVER_N0183_SERIAL, 0);
-      Nevent.SetPayload(payload);
-      m_parent_driver->AddPendingEvent(Nevent);
-    }
-
-    //  Handle pending output messages
-    std::string qmsg;
-    while (m_out_que.Get(qmsg)) {
-      if (-1 == WriteComPortPhysical(qmsg.c_str()) && 10 < m_send_retries++) {
-        // We failed to write the port 10 times, let's close the port so that
-        // the reconnection logic kicks in and tries to fix our connection.
-        m_send_retries = 0;
-        CloseComPortPhysical();
-      }
-    }
-  }
-
-  CloseComPortPhysical();
-  m_parent_driver->SetSecThreadInActive();  // I am dead
-  m_parent_driver->m_Thread_run_flag = -1;
-
-  return 0;
-}
-#endif  //__ANDROID__
